+++ src/lib/teemReporter.js
@@ -11,10 +11,10 @@
 const TeemDevice = require('@f5devcentral/f5-teem').Device;
 const TeemRecord = require('@f5devcentral/f5-teem').Record;
 
+const constants = require('./constants');
 const logger = require('./logger');
-const VERSION = require('./constants').VERSION;
-const APP_NAME = require('./constants').APP_NAME;
-const TELEMETRY_CONSUMER = require('./constants').CONFIG_CLASSES.CONSUMER_CLASS_NAME;
+
+const CONFIG_CLASSES = constants.CONFIG_CLASSES;
 
 /** @module TeemReporter */
 
@@ -28,8 +28,8 @@ const TELEMETRY_CONSUMER = require('./constants').CONFIG_CLASSES.CONSUMER_CLASS_
  */
 function TeemReporter() {
     this.assetInfo = {
-        name: APP_NAME,
-        version: VERSION
+        name: constants.APP_NAME,
+        version: constants.VERSION
     };
     this.teemDevice = new TeemDevice(this.assetInfo);
     this.logger = logger.getChild('teemReporter');
@@ -47,50 +47,94 @@ TeemReporter.prototype.process = function (config) {
 
     return Promise.resolve()
         .then(() => {
-            teemRecord = new TeemRecord(`${APP_NAME} Telemetry Data`, '1');
+            teemRecord = new TeemRecord(`${constants.APP_NAME} Telemetry Data`, '1');
         })
         .then(() => teemRecord.calculateAssetId())
         .then(() => teemRecord.addRegKey())
         .then(() => teemRecord.addPlatformInfo())
         .then(() => teemRecord.addProvisionedModules())
         .then(() => teemRecord.addClassCount(config))
-        .then(() => {
-            const extraFields = this._getCountByClassTypes(config, TELEMETRY_CONSUMER, 'consumers');
-            return teemRecord.addJsonObject(extraFields);
-        })
+        .then(() => teemRecord.addJsonObject(this.fetchExtraData(config)))
         .then(() => this.teemDevice.reportRecord(teemRecord))
         .catch(err => this.logger.exception('Unable to send analytics data', err));
 };
 
 /**
- * Count the number of types per class.
+ * Fetch extra data from the configuration
  *
- * @param {Object} declaration - the declaration to search through
- * @param {String} className - the declaration class to count
- * @param {String} keyName - the key to use when building count results
+ * @param {Object} config - the configuration to use to fetch extra data from
  *
- * @returns {Object} An object containing count results by type { keyName: { <type>: <count> } }
+ * @returns {Object} with extra data fetched from the configuration
  */
-TeemReporter.prototype._getCountByClassTypes = function (declaration, className, keyName) {
-    const results = { [keyName]: {} };
-    function getObjectWithClass(subDecl, cb) {
-        Object.keys(subDecl).forEach((key) => {
-            const val = subDecl[key];
-            if (typeof val === 'object') {
-                if (val.class === className) {
-                    cb(val);
-                }
-                getObjectWithClass(val, cb);
-            }
-        });
-    }
+TeemReporter.prototype.fetchExtraData = function (config) {
+    const extraData = {};
+    computeCounters(config, extraData);
+    return extraData;
+};
 
-    getObjectWithClass(declaration, (obj) => {
-        results[keyName][obj.type] = results[keyName][obj.type] ? results[keyName][obj.type] += 1 : 1;
+/**
+ * Calculate counters for Consumers (by type), inlined System Pollers and iHealth Pollers
+ *
+ * @param {Object} config    - the configuration to use to fetch extra data from
+ * @param {Object} extraData - object to write counters to
+ */
+function computeCounters(config, extraData) {
+    extraData = Object.assign(extraData, {
+        consumers: {},
+        inlineIHealthPollers: 0,
+        inlineSystemPollers: 0
     });
 
-    return results;
-};
+    const countConsumerType = (consumer) => {
+        extraData.consumers[consumer.type] = (extraData.consumers[consumer.type] || 0) + 1;
+    };
+
+    const countInlinePollers = (system) => {
+        if (typeof system.iHealthPoller === 'object') {
+            extraData.inlineIHealthPollers += 1;
+        }
+        if (Array.isArray(system.systemPoller)) {
+            system.systemPoller.forEach((systemPoller) => {
+                if (typeof systemPoller === 'object') {
+                    extraData.inlineSystemPollers += 1;
+                }
+            });
+        } else if (typeof system.systemPoller === 'object') {
+            extraData.inlineSystemPollers += 1;
+        }
+    };
+
+    searchObjectsWithClass(config, (obj) => {
+        switch (obj.class) {
+        case CONFIG_CLASSES.CONSUMER_CLASS_NAME:
+            countConsumerType(obj);
+            break;
+        case CONFIG_CLASSES.SYSTEM_CLASS_NAME:
+            countInlinePollers(obj);
+            break;
+        default:
+            // do nothing
+        }
+    });
+}
+
+/**
+ * Search for objects with 'class' property
+ *
+ * @param {Object} subDecl - declaration to traverse
+ * @param {Function} cb    - callback to call - cb(obj)
+ */
+function searchObjectsWithClass(subDecl, cb) {
+    Object.keys(subDecl).forEach((key) => {
+        const val = subDecl[key];
+        if (typeof val === 'object') {
+            if (typeof val.class === 'string') {
+                cb(val);
+            }
+            searchObjectsWithClass(val, cb);
+        }
+    });
+}
 
 module.exports = {
     TeemReporter