+++ src/lib/util.js
@@ -11,6 +11,7 @@
 const assignDefaults = require('lodash/defaultsDeep');
 const cloneDeep = require('lodash/cloneDeep');
 const clone = require('lodash/clone');
+const mergeWith = require('lodash/mergeWith');
 const fs = require('fs');
 const net = require('net');
 const path = require('path');
@@ -716,6 +717,47 @@ module.exports = {
         return cloneDeep(obj);
     },
 
+
+    /**
+     * Merges an Array of Objects into a single Object (uses lodash.mergeWith under the hood).
+     * Note: Nested Arrays are concatenated, not overwritten.
+     * Note: Passed data gets *spoiled* - copy if you need the original data
+     *
+     * @param {Array} collection - Array of Objects to be merged.
+     *
+     * @returns {Object} Object after being merged will all other Objects in the passed Array, or empty object
+     *
+     * Example:
+     *  collection: [
+     *      { a: 12, b: 13, c: [17, 18] },
+     *      { b: 14, c: [78, 79], d: 11 }
+     *  ]
+     *  will return:
+     *  {
+     *      a: 12, b: 14, c: [17, 18, 78, 79], d: 11
+     *  }
+     */
+    mergeObjectArray(collection) {
+        if (!Array.isArray(collection)) {
+            throw new Error('Expected input of Array');
+        }
+        // eslint-disable-next-line consistent-return
+        function concatArrays(objValue, srcValue) {
+            if (Array.isArray(objValue)) {
+                return objValue.concat(srcValue);
+            }
+        }
+
+        for (let i = 1; i < collection.length; i += 1) {
+            // only process elements that are Objects
+            if (typeof collection[i] === 'object' && !Array.isArray(collection[i])) {
+                mergeWith(collection[0], collection[i], concatArrays);
+            }
+        }
+        // First Object in array has been merged with all other Objects - it has the complete data set.
+        return collection[0] || {};
+    },
+
     /**
      * Start function
      *
@@ -791,16 +833,16 @@ module.exports = {
     },
 
     /**
-     * Stringify a message
+     * Stringify a message with option to pretty format
      *
      * @param {Object|String} msg - message to stringify
-     *
+     * @param {Boolean} prettyFormat - format JSON string to make it easier to read
      * @returns {Object} Stringified message
      */
-    stringify(msg) {
+    stringify(msg, pretty) {
         if (typeof msg === 'object') {
             try {
-                msg = JSON.stringify(msg);
+                msg = pretty ? JSON.stringify(msg, null, 4) : JSON.stringify(msg);
             } catch (e) {
                 // just leave original message intact
             }
@@ -896,6 +938,8 @@ module.exports = {
      *                                                   by default false
      * @param {Boolean} [options.includeResponseObject] - return [body, responseObject], by default false
      * @param {Array<Integer>|Integer} [options.expectedResponseCode]  - expected response code, by default 200
+     * @param {Integer} [options.timeout]               - Milliseconds to wait for a socket timeout. Option
+     *                                                    'passes through' to 'request' library
      *
      * @returns {Promise.<?any>} Returns promise resolved with response
      */