+++ test/unit/consumers/statsdConsumerTests.js
@@ -16,6 +16,7 @@ const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
 const proxyquire = require('proxyquire');
 const sinon = require('sinon');
+const net = require('net');
 
 const statsdExpectedData = require('./statsdConsumerTestsData');
 const testUtil = require('../shared/util');
@@ -27,30 +28,28 @@ describe('Statsd', () => {
     let statsDIndex; // later used to require ../../../src/lib/consumers/Statsd/index
 
     let passedClientParams;
-    let metrics;
+    let metrics = [];
 
     const defaultConsumerConfig = {
         host: 'statsd-host',
         port: '8125'
     };
 
+    const statsDClientStub = {
+        close: () => {},
+        gauge: (metricName, metricValue) => {
+            metrics.push({
+                metricName,
+                metricValue
+            });
+        }
+    };
+
     before(() => {
         statsDIndex = proxyquire('../../../src/lib/consumers/Statsd/index', {
-            'node-statsd': sinon.stub().callsFake((host, port) => {
-                passedClientParams = { host, port };
-                return {
-                    socket: {
-                        on: () => {}
-                    },
-                    close: () => {},
-                    gauge: (metricName, metricValue, cb) => {
-                        metrics.push({
-                            metricName,
-                            metricValue
-                        });
-                        cb(null, 'success');
-                    }
-                };
+            'statsd-client': sinon.stub().callsFake((options) => {
+                passedClientParams = { host: options.host, port: options.port, tcp: options.tcp };
+                return statsDClientStub;
             })
         });
     });
@@ -71,7 +70,33 @@ describe('Statsd', () => {
             return statsDIndex(context)
                 .then(() => assert.deepStrictEqual(passedClientParams, {
                     host: 'statsd-host',
-                    port: '8125'
+                    port: '8125',
+                    tcp: false
+                }));
+        });
+
+        it('should accept tcp as a protocol', () => {
+            sinon.stub(net, 'Socket').returns(
+                {
+                    on: () => {},
+                    connect: (port, host, cb) => { cb(); },
+                    end: () => { }
+                }
+            );
+            const context = testUtil.buildConsumerContext({
+                eventType: 'systemInfo',
+                config: {
+                    host: 'statsd-host',
+                    port: '8125',
+                    protocol: 'tcp'
+                }
+            });
+
+            return statsDIndex(context)
+                .then(() => assert.deepStrictEqual(passedClientParams, {
+                    host: 'statsd-host',
+                    port: '8125',
+                    tcp: true
                 }));
         });
 
@@ -94,6 +119,31 @@ describe('Statsd', () => {
 
             return statsDIndex(context)
                 .then(() => assert.deepStrictEqual(metrics, []));
+        });
+
+        it('should log exception if connection failure', () => {
+            sinon.stub(net, 'Socket').returns(
+                {
+                    on: (type, cb) => { cb('Connection failure to server'); },
+                    connect: () => { },
+                    destroy: () => { }
+                }
+            );
+            const context = testUtil.buildConsumerContext({
+                eventType: 'systemInfo',
+                config: {
+                    host: 'statsd-host',
+                    port: '8125',
+                    protocol: 'tcp'
+                }
+            });
+
+            return statsDIndex(context)
+                .then(() => {
+                    assert.strictEqual(context.logger.exception.callCount, 1);
+                    assert.deepEqual(context.logger.exception.firstCall.args,
+                        ['Unable to forward to statsd client', 'Connection failure to server']);
+                });
         });
     });
 });