+++ test/unit/shared/util.js
@@ -11,6 +11,7 @@
 const assert = require('assert');
 const cloneDeep = require('lodash/cloneDeep');
 const nock = require('nock');
+const sinon = require('sinon');
 
 const systemPollerData = require('../consumers/data/systemPollerData.json');
 const avrData = require('../consumers/data/avrData.json');
@@ -33,6 +34,34 @@ MockRestOperation.prototype.setStatusCode = function (code) { this.statusCode = 
 MockRestOperation.prototype.getUri = function () { return this.uri; };
 MockRestOperation.prototype.complete = function () { };
 
+/**
+ * Logger mock object
+ *
+ * @param {Object} [options]          - options when setting up logger mock
+ * @param {String} [options.logLevel] - log level to use. Default=info.
+ */
+function MockLogger(options) {
+    const opts = options || {};
+    this.logLevel = opts.logLevel || 'info';
+
+    this.setLogLevel = function (newLevel) {
+        this.logLevel = newLevel;
+    };
+    // Stubbed functions
+    this.error = sinon.stub();
+    this.debug = sinon.stub();
+    this.info = sinon.stub();
+    this.exception = sinon.stub();
+    // returns() returns the value passed at initialization - callsFake() uses CURRENT value
+    this.getLevelName = sinon.stub().callsFake(() => this.logLevel);
+}
+
+/**
+ * Tracer mock object
+ */
+function MockTracer() {
+    this.write = sinon.stub();
+}
 
 module.exports = {
     MockRestOperation,
@@ -51,9 +80,10 @@ module.exports = {
     /**
      * Generates a consumer config
      *
-     * @param {Object} [options]           - options to pass to context builder
-     * @param {String} [options.eventType] - which event type to return
-     * @param {Object} [options.config]    - optional config to inject into context
+     * @param {Object} [options]                     - options to pass to context builder
+     * @param {String} [options.eventType]           - which event type to return
+     * @param {Object} [options.config]              - optional config to inject into context
+     * @param {String} [options.loggerOpts.logLevel] - log level to initialize Logger with
      *
      * @returns {Object} Returns object representing consumer config
      */
@@ -77,13 +107,8 @@ module.exports = {
                 data,
                 type: opts.eventType || ''
             },
-            logger: {
-                error: () => { },
-                debug: () => { },
-                info: () => { },
-                exception: () => { },
-                getLevelName: () => { }
-            }
+            logger: new MockLogger(opts.loggerOpts),
+            tracer: new MockTracer()
         };
         return context;
     },